Moved analysis.py to correct location
