--- analysis/data/analysis.py
@@ -1,137 +1,0 @@
-"""
-
-Analysis.py - initial analyses for dhmit/rereading
-
-"""
-from ast import literal_eval
-import csv
-from pathlib import Path
-import unittest
-
-
-def load_data_csv(csv_path: Path):
-    """
-    Takes the path to a csv file, reads it, and returns its
-    content as a list of OrderedDicts
-    :param Path csv_path: path to the CSV file
-    :return: List[dict]
-    """
-    out_data = []
-    with open(str(csv_path)) as csv_file:
-        reader = csv.DictReader(csv_file)
-        for row in reader:
-            row['views'] = literal_eval(row['views'])
-            for header_name in ('id', 'student_id', 'scroll_ups'):
-                row[header_name] = int(row[header_name])
-            row = dict(row)
-            out_data.append(row)
-    return out_data
-
-
-def get_sentiments() -> dict:
-    """
-    Returns a dictionary of sentiment scores, with the keys being the word and the values being
-    their score
-
-    :return: dict mapping words to their sentiment scores
-    """
-    sentiment_path = Path('data', 'sentiments.txt')
-
-    sentiments = dict()
-    with open(sentiment_path, 'r') as file:
-        word = file.readline()
-
-        # We want to handle each word individually, rather than as a whole set
-        while word is not '':
-
-            # This particular file starts lines with '#' for non-sentiment comments, so skip them
-            if word[0] == '#' or word[0] == '\t':
-                word = file.readline()
-                continue
-
-            # All words use tabs to define the different parts of the data
-            attributes = word.split('\t')
-
-            # Pull out the word from the line
-            data = attributes[4]
-            data = data.split('#')
-            new_word = data[0]
-
-            # Find the largest sentiment score for the word, and define negative sentiments
-            # as negative values (if there's a tie, the sentiment is 0)
-            if attributes[2] == attributes[3]:
-                score = 0
-            elif attributes[2] > attributes[3]:
-                score = float(attributes[2])
-            else:
-                score = -float(attributes[3])
-
-            sentiments[new_word] = score
-
-            word = file.readline()
-
-    return sentiments
-
-
-def compute_total_view_time(student_data):
-    """
-    Given a list of student response dicts,
-    return the total time (across all users) spent reading the text
-
-    :param student_data: list, student response dicts
-    :return: float, the total time all users spent reading the text
-    """
-    total_view_time = 0
-    for row in student_data:
-        for view_time in row.get('views'):
-            total_view_time += view_time
-    return total_view_time
-
-
-def run_analysis():
-    """
-    Runs the analytical method on the reading data
-
-    :return: None
-    """
-    csv_path = Path('data', 'rereading_data_2019-09-13.csv')
-    student_data = load_data_csv(csv_path)
-
-    total_view_time = compute_total_view_time(student_data)
-    print(f'The total view time of all students was {total_view_time}.')
-
-
-class TestAnalysisMethods(unittest.TestCase):
-    """
-    Test cases to make sure things are running properly
-    """
-    def setUp(self):
-        test_data_path = Path('data', 'test_data.csv')
-        self.test_student_data = load_data_csv(test_data_path)
-        self.default_student_data = [  # model default values
-            {
-                'id': 0,
-                'question': '',
-                'context': '',
-                'response': '',
-                'views': [],
-                'student_id': 0,
-                'scroll_ups': 0,
-            }
-        ]
-
-    def test_compute_total_view_time(self):
-        """
-        Test that the total view time equals the expected values.
-        """
-        total_view_time = compute_total_view_time(self.test_student_data)
-        self.assertEqual(total_view_time, 6.385)
-
-        # check we don't crash on the defaults from the model!
-        total_view_time = compute_total_view_time(self.default_student_data)
-        self.assertEqual(total_view_time, 0)
-
-
-if __name__ == '__main__':
-    run_analysis()
-    unittest.main()  # run the tests
